% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_aa_clicked_to_ngl.R
\name{aa_clicked_to_ngl}
\alias{aa_clicked_to_ngl}
\title{Clicked residues to NGL query}
\usage{
aa_clicked_to_ngl(selected_aa, chainnames, selchain = NULL)
}
\arguments{
\item{selected_aa}{selected aa in list format}

\item{chainnames}{all chainnames of the sequence}

\item{selchain}{selected chainname}
}
\description{
Match clicked residues to chain and transform to NGL query
}
\examples{
\dontrun{
NGLVieweR_proxy("structure") \%>\% updateSelection("aa_clicked", 
sele = aa_clicked_to_ngl(aa_clicked(), structure_chainname(), selectedChain()))
}
}
