% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{bs_input_modal}
\alias{bs_input_modal}
\title{modal for bs_Inputs}
\usage{
bs_input_modal(id, title, body, size)
}
\arguments{
\item{id}{The id of the element. should be similar as id_modal given in the \code{bs_textInput()} function.}

\item{title}{Title of the modal.}

\item{body}{The body of the modal. Can also be a htmlTemplate.}

\item{size}{Size of the modal.}
}
\description{
Create a modal for a bs_input. To activate the modal it needs to be entered in the UI.
}
