% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\name{bs_textInput}
\alias{bs_textInput}
\title{textInput with bs info_modal}
\usage{
bs_textInput(id, label, placeholder = NULL, value = "", width = NULL, id_modal)
}
\arguments{
\item{id}{The id of the element}

\item{label}{label of textInput.}

\item{placeholder}{Input placeholder defaults to \code{NULL}}

\item{value}{value of textInput.}

\item{width}{width of textInput.}

\item{id_modal}{Id of the modal to show.}
}
\description{
Create a textInput with a bs info_modal.
}
