% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_insertUI_component.R
\name{insertUI_component}
\alias{insertUI_component}
\title{Insert components to UI}
\usage{
insertUI_component(type, name, uu_id)
}
\arguments{
\item{type}{type of components. E.g. label, selection, contact.}

\item{name}{name of the components. Use \code{insertUI_name()} to generate a unique name.}

\item{uu_id}{uniqueID used to name the UI components.}
}
\description{
Function to add UI components on submission
}
\examples{
\dontrun{
insertUI_component('selection', insertUI_name('ligand'), 
                   counter = 3, uu_id = "575ceb43-9ff5-40fb-85a4-9a8e9e06ceff")
}
}
