% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_insertUI_component.R
\name{insertUI_name}
\alias{insertUI_name}
\title{component UI name}
\usage{
insertUI_name(type, name = NULL, counter = 0)
}
\arguments{
\item{type}{type of selection. E.g. label, selection, contact.}

\item{name}{name of the selection. Defaults to selection-\code{counter} if input is \code{NULL} or \code{""}}

\item{counter}{value to create unique \code{name} if none is provided}
}
\description{
Function to add a unique name
}
\examples{
\dontrun{
insertUI_name('selection', '', 3)
"selection-3"
}
}
