% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_inputs.R
\docType{data}
\name{jsboxCollapse}
\alias{jsboxCollapse}
\title{Collapse box on title click}
\format{
An object of class \code{character} of length 1.
}
\usage{
jsboxCollapse
}
\description{
Collapse box on title click
}
\examples{
tags$head(  
extendShinyjs(text = jsboxCollapse, functions = c("collapse"))
)
observeEvent(input$sequenceTitle, {
js$collapse("sequence")
})
    
}
\keyword{datasets}
