% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadContacts.R
\name{loadContacts}
\alias{loadContacts}
\title{Load contacts from a .ngl file}
\usage{
loadContacts(
  NGLVieweR,
  contacts,
  contactTypes = list("hydrogenBond", "weakHydrogenBond", "waterHydrogenBond",
    "backboneHydrogenBond", "hydrophobic", "halogenBond", "ionicInteraction",
    "metalCoordination", "cationPi", "piStacking")
)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{contacts}{data.frame of contacts loaded from .ngl file.}

\item{contactTypes}{list of all different contactTypes that should be matched.}
}
\description{
Load contacts from a .ngl file
}
