% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadLabels.R
\name{loadLabels}
\alias{loadLabels}
\title{Load labels from a .ngl file}
\usage{
loadLabels(NGLVieweR, labels)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{labels}{data.frame of labels loaded from .ngl file.}
}
\description{
Load labels from a .ngl file
}
