% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadLigand.R
\name{loadLigand}
\alias{loadLigand}
\title{Load structure from a .ngl file}
\usage{
loadLigand(NGLVieweR, ligand)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{ligand}{data.frame of ligand loaded from .ngl file.}
}
\description{
Load structure from a .ngl file
}
