% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadSelections.R
\name{loadSelections}
\alias{loadSelections}
\title{Load selections from a .ngl file}
\usage{
loadSelections(NGLVieweR, selections)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{selections}{data.frame of selections loaded from .ngl file.}
}
\description{
Load selections from a .ngl file
}
