% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadStage.R
\name{loadStage}
\alias{loadStage}
\title{Load stage from a .ngl file}
\usage{
loadStage(NGLVieweR, stage)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{stage}{data.frame of selections loaded from .ngl file.}
}
\description{
Load stage from a .ngl file
}
