% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadStructure.R
\name{loadStructure}
\alias{loadStructure}
\title{Load structure from a .ngl file}
\usage{
loadStructure(NGLVieweR, structure, format = NULL)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{structure}{data.frame of structure loaded from .ngl file.}

\item{format}{File}
}
\description{
Load structure from a .ngl file
}
