% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadSurface.R
\name{loadSurface}
\alias{loadSurface}
\title{Load surface from a .ngl file}
\usage{
loadSurface(NGLVieweR, surface)
}
\arguments{
\item{NGLVieweR}{NGLVIeweR object.}

\item{surface}{data.frame of surface loaded from .ngl file.}
}
\description{
Load surface from a .ngl file
}
