% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_loadUI_component.R
\name{loadUI_component}
\alias{loadUI_component}
\title{Load UI components}
\usage{
loadUI_component(data, type)
}
\arguments{
\item{data}{data.frame of saved components}

\item{type}{type of component. E.g. label, selection, contact}
}
\description{
Function to load UI components
}
