% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ngl_to_position.R
\name{ngl_to_position}
\alias{ngl_to_position}
\title{AA positions from NGL query}
\usage{
ngl_to_position(sequence, input)
}
\arguments{
\item{sequence}{Protein sequence in string format}

\item{input}{sequence to find the positions for}
}
\description{
Find all AA positions in protein sequence from a NGL query. Returns a list of all matched positions.
}
\examples{
ngl_to_position("ALAAGSDFG", "1-5 OR <DFG>")
}
