% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readFile.R
\name{readFile}
\alias{readFile}
\title{Function to read structural files or NGL session}
\usage{
readFile(file)
}
\arguments{
\item{file}{The file input. Either a PDB code or structural file.}
}
\description{
Function to read .PDB or .NGL files
}
