% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sequence_df.R
\name{sequence_df}
\alias{sequence_df}
\title{Load sequence data.frame}
\usage{
sequence_df(sequence, resno, chainname, selchain = NULL)
}
\arguments{
\item{sequence}{sequence in list form.}

\item{resno}{amino acid numbers in list format}

\item{chainname}{chainname for each amino acid position in list format}

\item{selchain}{filter data for selected chain. Defaults to \code{NULL}}
}
\description{
Load sequence data.frame for D3 sequence input
}
